added apis

+++ src/main/java/com/clipboardhealth/summarystatsservice/pojo/view/EmployeeView.java
@@ -22,7 +22,7 @@ public class EmployeeView {
 
     private String name;
 
-    private Integer salary;
+    private Double salary;
 
     private Currency currency;
